convert multi-step-transform to ts
